[CARBONDATA-3641] Refactory data loading for partition table
[Background]

Currently, CarbonData only implemented hadoop commit algorithm version 1, which generated too many segment files during loading and generated too many small data files and index files

[Modification]
 1. implemented carbon commit algorithm, avoid to move data file and index files
 2. generate the final segment file directly
 3. optimize global_sort to avoid small files issue
 4. support complex data type in partition table (non-partition column)

This closes #3535

+++ core/src/main/java/org/apache/carbondata/core/util/CarbonUtil.java
@@ -338,9 +338,11 @@ public final class CarbonUtil {
       @Override
       public Void run() throws Exception {
         for (int i = 0; i < file.length; i++) {
-          boolean delete = file[i].delete();
-          if (!delete) {
-            throw new IOException("Error while deleting file: " + file[i].getAbsolutePath());
+          if (file[i].exists()) {
+            boolean delete = file[i].delete();
+            if (!delete) {
+              throw new IOException("Error while deleting file: " + file[i].getAbsolutePath());
+            }
           }
         }
         return null;
@@ -355,9 +357,11 @@ public final class CarbonUtil {
       @Override
       public Void run() throws Exception {
         for (int i = 0; i < file.length; i++) {
-          boolean delete = file[i].delete();
-          if (!delete) {
-            LOGGER.warn("Unable to delete file: " + file[i].getCanonicalPath());
+          if (file[i].exists()) {
+            boolean delete = file[i].delete();
+            if (!delete) {
+              LOGGER.warn("Unable to delete file: " + file[i].getCanonicalPath());
+            }
           }
         }
         return null;
@@ -2732,22 +2736,50 @@ public final class CarbonUtil {
     return Base64.decodeBase64(objectString.getBytes(CarbonCommonConstants.DEFAULT_CHARSET));
   }
 
+  public static void copyCarbonDataFileToCarbonStorePath(
+      String localFilePath,
+      String targetPath, long fileSizeInBytes,
+      OutputFilesInfoHolder outputFilesInfoHolder) throws CarbonDataWriterException {
+    if (targetPath.endsWith(".tmp") && localFilePath
+        .endsWith(CarbonCommonConstants.FACT_FILE_EXT)) {
+      // for partition case, write carbondata file directly to final path, keep index in temp path.
+      // This can improve the commit job performance on s3a.
+      targetPath =
+          targetPath.substring(0, targetPath.lastIndexOf("/"));
+      if (outputFilesInfoHolder != null) {
+        outputFilesInfoHolder.addToPartitionPath(targetPath);
+      }
+    }
+    long targetSize = copyCarbonDataFileToCarbonStorePath(localFilePath, targetPath,
+        fileSizeInBytes);
+    if (outputFilesInfoHolder != null) {
+      // Storing the number of files written by each task.
+      outputFilesInfoHolder.incrementCount();
+      // Storing the files written by each task.
+      outputFilesInfoHolder.addToOutputFiles(targetPath + localFilePath
+          .substring(localFilePath.lastIndexOf(File.separator)) + ":" + targetSize);
+    }
+  }
+
   /**
    * This method will copy the given file to carbon store location
    *
    * @param localFilePath local file name with full path
    * @throws CarbonDataWriterException
+   * @return the file size
    */
-  public static void copyCarbonDataFileToCarbonStorePath(String localFilePath,
+  public static long copyCarbonDataFileToCarbonStorePath(String localFilePath,
       String carbonDataDirectoryPath, long fileSizeInBytes)
       throws CarbonDataWriterException {
     long copyStartTime = System.currentTimeMillis();
     LOGGER.info(String.format("Copying %s to %s, operation id %d", localFilePath,
         carbonDataDirectoryPath, copyStartTime));
+    long targetSize = 0;
     try {
       CarbonFile localCarbonFile = FileFactory.getCarbonFile(localFilePath);
+      long localFileSize = localCarbonFile.getSize();
       // the size of local carbon file must be greater than 0
-      if (localCarbonFile.getSize() == 0L) {
+      if (localFileSize == 0L) {
         LOGGER.error("The size of local carbon file: " + localFilePath + " is 0.");
         throw new CarbonDataWriterException("The size of local carbon file is 0.");
       }
@@ -2755,16 +2787,16 @@ public final class CarbonUtil {
           .substring(localFilePath.lastIndexOf(File.separator));
       copyLocalFileToCarbonStore(carbonFilePath, localFilePath,
           CarbonCommonConstants.BYTEBUFFER_SIZE,
-          getMaxOfBlockAndFileSize(fileSizeInBytes, localCarbonFile.getSize()));
+          getMaxOfBlockAndFileSize(fileSizeInBytes, localFileSize));
       CarbonFile targetCarbonFile = FileFactory.getCarbonFile(carbonFilePath);
       // the size of carbon file must be greater than 0
       // and the same as the size of local carbon file
-      if (targetCarbonFile.getSize() == 0L ||
-          (targetCarbonFile.getSize() != localCarbonFile.getSize())) {
+      targetSize = targetCarbonFile.getSize();
+      if (targetSize == 0L || targetSize != localFileSize) {
         LOGGER.error("The size of carbon file: " + carbonFilePath + " is 0 "
             + "or is not the same as the size of local carbon file: ("
-            + "carbon file size=" + targetCarbonFile.getSize()
-            + ", local carbon file size=" + localCarbonFile.getSize() + ")");
+            + "carbon file size=" + targetSize
+            + ", local carbon file size=" + localFileSize + ")");
         throw new CarbonDataWriterException("The size of carbon file is 0 "
             + "or is not the same as the size of local carbon file.");
       }
@@ -2774,6 +2806,7 @@ public final class CarbonUtil {
     }
     LOGGER.info(String.format("Total copy time is %d ms, operation id %d",
         System.currentTimeMillis() - copyStartTime, copyStartTime));
+    return targetSize;
   }
 
   /**
